pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mahjong-scoreboard"
version = "0.4.0a1"
description = "日麻寄分器（NoneBot插件）"
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.MD"
license = "MIT"
repository = "https://github.com/ssttkkl/nonebot-plugin-mahjong-scoreboard"
packages = [
    { include = "nonebot_plugin_mahjong_scoreboard", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0rc4"
nonebot-adapter-onebot = "^2.2.2"
nonebot_plugin_apscheduler = "^0.2.0"
nonebot-plugin-sqlalchemy = "^0.2.1"
nonebot-plugin-session = "^0.0.3"
nonebot-plugin-gocqhttp-cross-machine-upload-file = "^0.1.5"
aiosqlite = ">=0.17,<0.19"
tzlocal = "^4.2"
cachetools = "^5.2.0"
nonebot-plugin-localstore = "^0.4.1"

[tool.poetry.group.dev.dependencies]
nonebot2 = { extras = ["httpx", "fastapi"], version = "^2.0.0rc3" }
nonebot-adapter-qqguild = "^0.2.2"
nonebug = "^0.2.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
